chore: minor backward compatible refactor
